Fix potential empty chunks from make_chunks
make_chunks() could potentially yield a final chunk
containing a copy of init, with no additional items,
when the number of items equals the size.
This causes an indefinite hang when combined with the
exception-handling bug fixed in the preceding commit,
fde7c7199de6a8b028670714a0daa1c8fb327b9f
It appears the check for the final chunk size was written
before optional init values were added to the function.