Fix deletion of stored photo in photomanager clean
If a photo was indexed and collected at a lower priority
and indexed again at a higher priority, it will be marked
as stored by two photos. Running photomanager clean
would delete the photo as its store_path was listed in the
lower-priority PhotoFile. Instead, just de-list the stored photo
from the lower-priority PhotoFile while keeping it in the
higher-priority PhotoFile.