Fix hang in AsyncExifTool.do_job()
exiftool outputs a sentinel string `b"{ready}"` when it is done
outputting metadata for a given batch. Previously the code would read
strings of a certain block_size from exiftool's stdout in a while loop
that breaks when the sentinel is detected in the last block. However, if
the sentinel was split across two blocks, the program would hang
indefinitely.

Instead of checking the last block, check the concatenation of the last
two blocks for the sentinel string.